Criando uma classe de erros personalizada

+++ src/storage/group/groupCreate.ts
@@ -1,4 +1,6 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
+import { AppError } from "@utils/AppError";
+
 import { groupsGetAll } from "./groupsGetAll";
 import { GROUP_COLLECTION } from "@storage/storageConfig";
 
@@ -7,6 +9,12 @@ export async function groupCreate(newGroup: string) {
     const storedGroups = await groupsGetAll();
 
     const storage = JSON.stringify([...storedGroups, newGroup]);
+
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if (groupAlreadyExists) {
+      throw new AppError("Já existe um grupo cadastrado com esse nome.");
+    }
 
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
   } catch (error) {
